+++ aerich/ddl/mysql/__init__.py
@@ -1,4 +1,6 @@
-from typing import TYPE_CHECKING, List, Type
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
 
 from tortoise.backends.mysql.schema_generator import MySQLSchemaGenerator
 
@@ -21,9 +23,7 @@ class MysqlDDL(BaseDDL):
     _RENAME_COLUMN_TEMPLATE = (
         "ALTER TABLE `{table_name}` RENAME COLUMN `{old_column_name}` TO `{new_column_name}`"
     )
-    _ADD_INDEX_TEMPLATE = (
-        "ALTER TABLE `{table_name}` ADD {unique}INDEX `{index_name}` ({column_names})"
-    )
+    _ADD_INDEX_TEMPLATE = "ALTER TABLE `{table_name}` ADD {index_type}{unique}INDEX `{index_name}` ({column_names}){extra}"
     _DROP_INDEX_TEMPLATE = "ALTER TABLE `{table_name}` DROP INDEX `{index_name}`"
     _ADD_FK_TEMPLATE = "ALTER TABLE `{table_name}` ADD CONSTRAINT `{fk_name}` FOREIGN KEY (`{db_column}`) REFERENCES `{table}` (`{field}`) ON DELETE {on_delete}"
     _DROP_FK_TEMPLATE = "ALTER TABLE `{table_name}` DROP FOREIGN KEY `{fk_name}`"
@@ -36,7 +36,7 @@ class MysqlDDL(BaseDDL):
     _MODIFY_COLUMN_TEMPLATE = "ALTER TABLE `{table_name}` MODIFY COLUMN {column}"
     _RENAME_TABLE_TEMPLATE = "ALTER TABLE `{old_table_name}` RENAME TO `{new_table_name}`"
 
-    def _index_name(self, unique: bool, model: "Type[Model]", field_names: List[str]) -> str:
+    def _index_name(self, unique: bool | None, model: type[Model], field_names: list[str]) -> str:
         if unique:
             if len(field_names) == 1:
                 # Example: `email = CharField(max_length=50, unique=True)`
@@ -47,17 +47,3 @@ class MysqlDDL(BaseDDL):
         else:
             index_prefix = "idx"
         return self.schema_generator._generate_index_name(index_prefix, model, field_names)
-
-    def add_index(self, model: "Type[Model]", field_names: List[str], unique=False) -> str:
-        return self._ADD_INDEX_TEMPLATE.format(
-            unique="UNIQUE " if unique else "",
-            index_name=self._index_name(unique, model, field_names),
-            table_name=model._meta.db_table,
-            column_names=", ".join(self.schema_generator.quote(f) for f in field_names),
-        )
-
-    def drop_index(self, model: "Type[Model]", field_names: List[str], unique=False) -> str:
-        return self._DROP_INDEX_TEMPLATE.format(
-            index_name=self._index_name(unique, model, field_names),
-            table_name=model._meta.db_table,
-        )