+++ aerich/ddl/postgres/__init__.py
@@ -1,4 +1,6 @@
-from typing import Type, cast
+from __future__ import annotations
+
+from typing import cast
 
 from tortoise import Model
 from tortoise.backends.asyncpg.schema_generator import AsyncpgSchemaGenerator
@@ -9,7 +11,7 @@ from aerich.ddl import BaseDDL
 class PostgresDDL(BaseDDL):
     schema_generator_cls = AsyncpgSchemaGenerator
     DIALECT = AsyncpgSchemaGenerator.DIALECT
-    _ADD_INDEX_TEMPLATE = 'CREATE {unique}INDEX "{index_name}" ON "{table_name}" ({column_names})'
+    _ADD_INDEX_TEMPLATE = 'CREATE {unique}INDEX IF NOT EXISTS "{index_name}" ON "{table_name}" {index_type}({column_names}){extra}'
     _DROP_INDEX_TEMPLATE = 'DROP INDEX IF EXISTS "{index_name}"'
     _ALTER_NULL_TEMPLATE = 'ALTER TABLE "{table_name}" ALTER COLUMN "{column}" {set_drop} NOT NULL'
     _MODIFY_COLUMN_TEMPLATE = (
@@ -18,7 +20,7 @@ class PostgresDDL(BaseDDL):
     _SET_COMMENT_TEMPLATE = 'COMMENT ON COLUMN "{table_name}"."{column}" IS {comment}'
     _DROP_FK_TEMPLATE = 'ALTER TABLE "{table_name}" DROP CONSTRAINT IF EXISTS "{fk_name}"'
 
-    def alter_column_null(self, model: "Type[Model]", field_describe: dict) -> str:
+    def alter_column_null(self, model: type[Model], field_describe: dict) -> str:
         db_table = model._meta.db_table
         return self._ALTER_NULL_TEMPLATE.format(
             table_name=db_table,
@@ -26,7 +28,7 @@ class PostgresDDL(BaseDDL):
             set_drop="DROP" if field_describe.get("nullable") else "SET",
         )
 
-    def modify_column(self, model: "Type[Model]", field_describe: dict, is_pk: bool = False) -> str:
+    def modify_column(self, model: type[Model], field_describe: dict, is_pk: bool = False) -> str:
         db_table = model._meta.db_table
         db_field_types = cast(dict, field_describe.get("db_field_types"))
         db_column = field_describe.get("db_column")
@@ -38,7 +40,7 @@ class PostgresDDL(BaseDDL):
             using=f' USING "{db_column}"::{datatype}',
         )
 
-    def set_comment(self, model: "Type[Model]", field_describe: dict) -> str:
+    def set_comment(self, model: type[Model], field_describe: dict) -> str:
         db_table = model._meta.db_table
         return self._SET_COMMENT_TEMPLATE.format(
             table_name=db_table,