+++ tests/test_ddl.py
@@ -1,3 +1,5 @@
+import tortoise
+
 from aerich.ddl.mysql import MysqlDDL
 from aerich.ddl.postgres import PostgresDDL
 from aerich.ddl.sqlite import SqliteDDL
@@ -8,6 +10,21 @@ from tests.models import Category, Product, User
 def test_create_table():
     ret = Migrate.ddl.create_table(Category)
     if isinstance(Migrate.ddl, MysqlDDL):
+        if tortoise.__version__ >= "0.24":
+            assert (
+                ret
+                == """CREATE TABLE IF NOT EXISTS `category` (
+    `id` INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
+    `slug` VARCHAR(100) NOT NULL,
+    `name` VARCHAR(200),
+    `title` VARCHAR(20) NOT NULL,
+    `created_at` DATETIME(6) NOT NULL DEFAULT CURRENT_TIMESTAMP(6),
+    `owner_id` INT NOT NULL COMMENT 'User',
+    CONSTRAINT `fk_category_user_110d4c63` FOREIGN KEY (`owner_id`) REFERENCES `user` (`id`) ON DELETE CASCADE,
+    FULLTEXT KEY `idx_category_slug_e9bcff` (`slug`)
+) CHARACTER SET utf8mb4"""
+            )
+            return
         assert (
             ret
             == """CREATE TABLE IF NOT EXISTS `category` (
@@ -18,20 +35,23 @@ def test_create_table():
     `created_at` DATETIME(6) NOT NULL DEFAULT CURRENT_TIMESTAMP(6),
     `owner_id` INT NOT NULL COMMENT 'User',
     CONSTRAINT `fk_category_user_110d4c63` FOREIGN KEY (`owner_id`) REFERENCES `user` (`id`) ON DELETE CASCADE
-) CHARACTER SET utf8mb4"""
+) CHARACTER SET utf8mb4;
+CREATE FULLTEXT INDEX `idx_category_slug_e9bcff` ON `category` (`slug`)"""
         )
 
     elif isinstance(Migrate.ddl, SqliteDDL):
+        exists = "IF NOT EXISTS " if tortoise.__version__ >= "0.24" else ""
         assert (
             ret
-            == """CREATE TABLE IF NOT EXISTS "category" (
+            == f"""CREATE TABLE IF NOT EXISTS "category" (
     "id" INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
     "slug" VARCHAR(100) NOT NULL,
     "name" VARCHAR(200),
     "title" VARCHAR(20) NOT NULL,
     "created_at" TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "owner_id" INT NOT NULL REFERENCES "user" ("id") ON DELETE CASCADE /* User */
-)"""
+);
+CREATE INDEX {exists}"idx_category_slug_e9bcff" ON "category" ("slug")"""
         )
 
     elif isinstance(Migrate.ddl, PostgresDDL):
@@ -45,6 +65,7 @@ def test_create_table():
     "created_at" TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "owner_id" INT NOT NULL REFERENCES "user" ("id") ON DELETE CASCADE
 );
+CREATE INDEX IF NOT EXISTS "idx_category_slug_e9bcff" ON "category" USING HASH ("slug");
 COMMENT ON COLUMN "category"."owner_id" IS 'User'"""
         )
 
@@ -163,6 +184,14 @@ def test_add_index():
     if isinstance(Migrate.ddl, MysqlDDL):
         assert index == "ALTER TABLE `category` ADD INDEX `idx_category_name_8b0cb9` (`name`)"
         assert index_u == "ALTER TABLE `category` ADD UNIQUE INDEX `name` (`name`)"
+    elif isinstance(Migrate.ddl, PostgresDDL):
+        assert (
+            index == 'CREATE INDEX IF NOT EXISTS "idx_category_name_8b0cb9" ON "category" ("name")'
+        )
+        assert (
+            index_u
+            == 'CREATE UNIQUE INDEX IF NOT EXISTS "uid_category_name_8b0cb9" ON "category" ("name")'
+        )
     else:
         assert index == 'CREATE INDEX "idx_category_name_8b0cb9" ON "category" ("name")'
         assert index_u == 'CREATE UNIQUE INDEX "uid_category_name_8b0cb9" ON "category" ("name")'